requirements.txt
google-cloud-vision==3.4.4
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
jinja2==3.1.2
